Validation branch

+++ src/app/components/tasks/tasks.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import {ApiService} from '../../service/api.service';
+import {AuthService} from '../../service/auth.service';
 
 @Component({
   selector: 'app-tasks',
@@ -11,7 +12,8 @@ export class TasksComponent implements OnInit {
   tasks = [];
   todo_id: number;
   task;
-  constructor(private route: ActivatedRoute, private apiService: ApiService ) { }
+  can;
+  constructor(private route: ActivatedRoute, private apiService: ApiService, private authService: AuthService) { }
 
   ngOnInit() {
     this.route.paramMap.subscribe(params => {
@@ -20,11 +22,12 @@ export class TasksComponent implements OnInit {
           console.log(tasks);
           this.tasks = tasks;
         });
-        this.apiService.getTodo(this.todo_id).subscribe( task=>{
-          this.task = task;
-          console.log(task);
-        });
+        // this.apiService.getTodo(this.todo_id).subscribe( task=> {
+        //   this.task = task;
+        //   console.log(task);
+        // });
     });
+    this.canAdd();
   }
 
 
@@ -35,17 +38,37 @@ export class TasksComponent implements OnInit {
         console.log(tasks);
         this.tasks = tasks;
       });
-    })
+    });
   }
 
   public onItemDelete(todo_id, item_id) {
-    this.apiService.deleteItem(todo_id, item_id).subscribe(task=> {
-      this.apiService.getallTasks(this.todo_id).subscribe(tasks =>{
-        console.log(tasks);
-        this.tasks = tasks;
+    const r = confirm('Are you sure to delete this item?');
+    if (r) {
+      this.apiService.deleteItem(todo_id, item_id).subscribe(task=> {
+        this.apiService.getallTasks(this.todo_id).subscribe(tasks =>{
+          console.log(tasks);
+          this.tasks = tasks;
+        });
       });
-    });
-
+    } else {
+      window.alert('Action reverted');
+    }
+  }
+  public canAdd() {
+    if (this.authService.isLoggedIn()) {
+      const id = this.authService.getuserId();
+      this.apiService.getTodo(this.todo_id).subscribe(todo=>{
+        console.log('todo: ', todo);
+        if (id == todo.created_by) {
+          console.log('true');
+          this.can = true;
+        } else {
+          this.can = false;
+        }
+      });
+    } else {
+      this.can = false;
+    }
   }
 
 }
